Struct nphysics::detection::joint::Fixed

pub struct Fixed {
    // some fields omitted
}

A joint that prevents any relative movement (linear and angular) between two objects.

Methods

impl Fixed

fn new(anchor1: Anchor<Matrix>, anchor2: Anchor<Matrix>) -> Fixed

Creates a new Fixed joint.

fn up_to_date(&self) -> bool

Tells if the joint has been modified by the user.

fn set_local1(&mut self, local1: Matrix)

Sets the the second anchor position.

The position is expressed in the second attached body’s local coordinates.

fn set_local2(&mut self, local2: Matrix)

Sets the the second anchor position.

The position is expressed in the second attached body’s local coordinates.