Struct nphysics::detection::joint::JointManager

pub struct JointManager {
    // some fields omitted
}

Structure that handles creation and removal of joints.

Methods

impl JointManager

fn new() -> JointManager

Creates a new JointManager.

fn joints(&self) -> &HashMap<usize, Constraint, UintTWHash>

Joints handled by this manager.

fn joints_with_body(&self, body: &Rc<RefCell<RigidBody>>) -> Option<&[Constraint]>

List of joints attached to a specific body.

fn add_ball_in_socket(&mut self, joint: Rc<RefCell<BallInSocket>>, activation: &mut ActivationManager)

Add a BallInSocket joint to this manager.

This will force the activation of the two objects attached to the joint.

fn remove_ball_in_socket(&mut self, joint: &Rc<RefCell<BallInSocket>>, activation: &mut ActivationManager)

Removes a BallInSocket joint from this manager.

This will force the activation of the two objects attached to the joint.

fn add_fixed(&mut self, joint: Rc<RefCell<Fixed>>, activation: &mut ActivationManager)

Add a Fixed joint to this manager.

This will force the activation of the two objects attached to the joint.

fn remove_joint<T: Joint<M>, M>(&mut self, joint: &Rc<RefCell<T>>, activation: &mut ActivationManager)

Removes a joint from this manager.

This will force the activation of the two objects attached to the joint.

fn remove(&mut self, b: &Rc<RefCell<RigidBody>>, activation: &mut ActivationManager)

Removes every joint attached to a given rigid body.

This will force the activation of every object attached to the deleted joints.

fn update(&mut self, activation: &mut ActivationManager)

Activates the objects that interact with an activated object through a joint.

fn interferences(&mut self, constraint: &mut Vec<Constraint>)

Collects all the constraints caused by joints.