Struct nphysics2df64​::detection​::BodiesBodies​

pub struct BodiesBodies<BF> {
    // some fields omitted
}

Collision detector between rigid bodies.

Methods

impl<BF: InterferencesBroadPhase<Rc<RefCell<RigidBody>>, Box<GeomGeomCollisionDetector>>> BodiesBodies<BF>

fn new(dispatcher: Rc<GeomGeomDispatcher>) -> BodiesBodies<BF>

Creates a new BodiesBodies collision detector.

impl<BF: RayCastBroadPhase<Rc<RefCell<RigidBody>>>> BodiesBodies<BF>

fn interferences_with_ray(&mut self, ray: &Ray, broad_phase: &mut BF, out: &mut Vec<(Rc<RefCell<RigidBody>>, Scalar)>)

Computes the interferences between every rigid bodies of a given broad phase, and a ray.

impl<BF: BoundingVolumeBroadPhase<Rc<RefCell<RigidBody>>, AABB>> BodiesBodies<BF>

fn remove(&mut self, o: &Rc<RefCell<RigidBody>>, broad_phase: &mut BF, activation: &mut ActivationManager)

Removes a rigid body from this detector.

This must be called whenever a rigid body is removed from the physics world.