Function nphysics2df64​::integration​::euler​::displacement​

pub fn displacement(dt: Scalar, _: &Matrix, center_of_mass: &Vect, lin_vel: &Vect, ang_vel: &Orientation) -> Matrix

Computes the transformation matrix required to move an object with a lin_vel linear velocity, a ang_vil angular velocity, and a center of mass center_of_mass, during the time step dt.