Struct nphysics::resolution::AccumulatedImpulseSolver

pub struct AccumulatedImpulseSolver {
    // some fields omitted
}

Constraint solver using the projected gauss seidel algorithm and warm-starting.

Methods

impl AccumulatedImpulseSolver

fn new(step: Scalar, correction_mode: CorrectionMode, joint_corr_factor: Scalar, rest_eps: Scalar, num_first_order_iter: usize, num_second_order_iter: usize) -> AccumulatedImpulseSolver

Creates a new AccumulatedImpulseSolver.

fn num_first_order_iter(&self) -> usize

Gets the number of iteration done by the penetration depth correction solver.

fn set_num_first_order_iter(&mut self, num: usize)

Sets the number of iteration done by the penetration depth correction solver.

fn num_second_order_iter(&self) -> usize

Gets the number of iteration done by the velocity constraint solver.

fn set_num_second_order_iter(&mut self, num: usize)

Sets the number of iteration done by the velocity constraint solver.